test(infra): split NUL-tail recovery tests to satisfy max-lines
